changes the  receivedTime to float64

+++ waku/node/v2/waku_types.nim
@@ -39,7 +39,7 @@ type
   Index* = object
     ## This type contains the  description of an index used in the pagination of waku messages
     digest*: MDigest[256]
-    receivedTime*: float
+    receivedTime*: float64
 
   IndexedWakuMessage* = object
     msg*: WakuMessage
